+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@ ones are marked like "v1.0.0-fork".
 
 ### Added
 
+* Support for Japanese with MeCab on Mac! This was added thanks to 
+[quopquai](https://github.com/quopquai) on [#135](https://github.com/HugoFara/lwt/issues/135).
 * `unloadformcheck.js` now declares a new object `lwt_form_check` that contains all the functions needed.
 * New globals ([#163](https://github.com/HugoFara/lwt/issues/163)):
   * On `inc/kernel_utility.php`: `LWT_APP_VERSION` and `LWT_RELEASE_DATE`.